requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
redis
Flask-JWT-Extended
passlib[bcrypt]
